requirements.txt
# Runtime dependencies
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.8
python-dotenv==1.0.0
flask_cors

# Runtime tools
gunicorn==20.1.0

# Code quality
pylint==2.17.4
flake8==5.0.4
black==23.3.0

# Testing dependencies
nose==1.3.7
pytest
pinocchio==0.4.3
factory-boy==3.2.1
coverage==7.1.0
httpie==3.2.1

# API documentation
flask-swagger-ui